requirements.txt
# Core Dependencies
python-telegram-bot==20.8
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2

# Database
sqlalchemy==2.0.23
python-dotenv==1.0.0

# Web Framework (for admin panel)
fastapi==0.104.1
uvicorn==0.24.0

# Payment Processing
stripe==7.5.0

# Scheduling
apscheduler==3.10.4

# Development Tools
pytest==7.4.3
black==23.11.0
flake8==6.1.0
